+++ components/server/src/data_model/metrics.py
@@ -179,7 +179,7 @@ METRICS = Metrics.parse_obj(
             unit=Unit.TRANSACTIONS,
             near_target="5",
             default_source="jmeter_json",
-            sources=["manual_number", "jmeter_json", "performancetest_runner"],
+            sources=["manual_number", "jmeter_csv", "jmeter_json", "performancetest_runner"],
             tags=[Tag.PERFORMANCE],
         ),
         source_up_to_dateness=dict(